Changed Game to HashMap of games and added private openings.

+++ src/main/java/chess_debiut/opening/OpeningService.java
@@ -1,20 +1,29 @@
 package chess_debiut.opening;
 
+import chess_debiut.user.User;
+import chess_debiut.user.UserRepository;
+import chess_debiut.user_opening.UserOpening;
+import chess_debiut.user_opening.UserOpeningRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-import java.util.Objects;
-import java.util.Optional;
+import java.time.LocalDate;
+import java.util.*;
 
 @Service
 public class OpeningService {
     private final OpeningRepository openingRepository;
 
+    private final UserOpeningRepository userOpeningRepository;
+
+    private final UserRepository userRepository;
     @Autowired
-    public OpeningService(OpeningRepository openingRepository) {
+    public OpeningService(OpeningRepository openingRepository, UserOpeningRepository userOpeningRepository, UserRepository userRepository) {
         this.openingRepository = openingRepository;
+        this.userOpeningRepository = userOpeningRepository;
+        this.userRepository = userRepository;
     }
 
     public List<Opening> getAllOpenings() {
@@ -22,7 +31,16 @@ public class OpeningService {
     }
 
     public void addNewOpening(Opening opening) {
+        LocalDate date = LocalDate.now();
+        opening.setCreationDate(date);
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        User currentUser = userRepository.getUserByUsername(username).get();
+        opening.setCreatedBy(currentUser);
         openingRepository.save(opening);
+        UserOpening userOpening = new UserOpening();
+        userOpening.setOpening(opening);
+        userOpening.setUser(currentUser);
+        userOpeningRepository.save(userOpening);
     }
 
     public void deleteOpening(Long openingId) {
@@ -31,7 +49,13 @@ public class OpeningService {
                     "Opening with id " + openingId + " does not exist."
             );
         }
-        openingRepository.deleteById(openingId);
+        Opening opening = openingRepository.findOpeningById(openingId).get();
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        if (opening.getCreatedBy().getUsername().equals(username)) {
+            openingRepository.deleteById(openingId);
+        } else {
+            throw new IllegalStateException("You can't delete opening that you didn't create.");
+        }
     }
 
     @Transactional
@@ -63,5 +87,50 @@ public class OpeningService {
         Optional<Opening> openingStartingWith = openingRepository.findOpeningStartingWith(nameFrag);
         return openingStartingWith.map(List::of).orElse(List.of());
 
+    }
+
+    public List<Opening> getUserOpenings() {
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        User currentUser = userRepository.getUserByUsername(username).get();
+        List<UserOpening> userSubscriptions = userOpeningRepository.findByUser(currentUser);
+        List<Opening> openings = new ArrayList<>();
+        for (UserOpening userOpening: userSubscriptions){
+            openings.add(userOpening.getOpening());
+        }
+        return openings;
+    }
+
+    public Optional<Opening> findOpeningById(Long openingID){
+        return openingRepository.findOpeningById(openingID);
+    }
+
+    public void subscribeOpening(Long openingId) {
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        User currentUser = userRepository.getUserByUsername(username).get();
+        if (!openingRepository.existsById(openingId)){
+            throw new IllegalStateException(
+                    "Opening with id " + openingId + " does not exist.");
+        }
+        Opening opening = openingRepository.findOpeningById(openingId).get();
+        UserOpening userOpening = new UserOpening();
+        userOpening.setUser(currentUser);
+        userOpening.setOpening(opening);
+        userOpeningRepository.save(userOpening);
+    }
+
+    public void unsubOpening(Long openingId) {
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        User currentUser = userRepository.getUserByUsername(username).get();
+        List<UserOpening> userSubscriptions = userOpeningRepository.findByUser(currentUser);
+        Integer relationToDeleteID = null;
+        for (UserOpening userOpening : userSubscriptions){
+            if (userOpening.getOpening().getId().equals(openingId)){
+                relationToDeleteID = userOpening.getID();
+                break;
+            }
+        }
+        if (relationToDeleteID != null) {
+            userOpeningRepository.deleteById(relationToDeleteID);
+        }
     }
 }
